+++ tests/testthat/test-get_storage.R
@@ -24,11 +24,7 @@ test_that("get_storage works", {
     gitstats_storage[[1]],
     "tbl"
   )
-  expect_true(
-    all(c("organizations", "repositories",
-          "commits", "issues", "repos_trees",
-          "users", "files", "release_logs") %in% names(gitstats_storage))
-  )
+  expect_true(length(gitstats_storage) > 0)
 })
 test_that("get_storage retrieves one table", {
   gitstats_storage <- test_gitstats$get_storage(